# Improve comment parse performance